requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
contextlib2==21.6.0
deepspeed==0.5.10
fairscale==0.4.5
flatbuffers==2.0
frozenlist==1.3.0
fsspec==2022.1.0
future==0.18.2
gast==0.5.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
hjson==3.0.2
idna==3.3
importlib-metadata==4.10.1
keras==2.8.0
Keras-Preprocessing==1.1.2
libclang==13.0.0
Markdown==3.3.6
MedPy==0.4.0
ml-collections==0.1.1
multidict==5.2.0
ninja==1.10.2.3
numpy==1.22.1
oauthlib==3.1.1
onnx==1.11.0
onnx-tf==1.9.0
onnxruntime==1.10.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
Pillow==9.0.0
plotly==5.5.0
protobuf==3.6.1
psutil==5.9.0
py-cpuinfo==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
pyparsing==3.0.7
pytorch-extension==0.2
pytorch-lightning==1.5.9
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
SimpleITK==2.1.1
six==1.16.0
tenacity==8.0.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.16.1
tensorflow-cpu==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
thop==0.0.31.post2005241907
torch==1.10.0+cu113
torchaudio==0.10.0+cu113
torchmetrics==0.7.0
torchsummary==1.5.1
torchvision==0.11.1+cu113
tqdm==4.62.3
triton==1.0.0
typeguard==2.13.3
typing==3.7.4.3
typing-extensions==4.0.1
urllib3==1.26.8
Werkzeug==2.0.2
wrapt==1.13.3
yacs==0.1.8
yarl==1.7.2
zipp==3.7.0
